fix: 1.fix the issue of failing to retrieve window ID
When using xdotool to retrieve window IDs, if there are multiple windows, the output ends with '\n'. In the original code, directly splitting the string using split("\n") can result in the last element of the resulting list being an empty string. This situation leads to errors when iterating through the window IDs and converting them to int type

Log: fix the issue of failing to retrieve window ID